test(ask-karma): split connection-drop tests into separate file
Move the DEV-394 mid-stream connection-drop tests (+ the dropped-stream
mock helper) out of useAgentStream.test.ts into a focused
useAgentStream.network-errors.test.ts. The original file was already over
the per-file size limit; adding the new cases there tripped the quality
gate's "grew while over limit" regression. Splitting restores the
original file to its prior size and keeps the new suite well under the
limit. No behavior or coverage change — same 51 tests pass.
